Proposed Cheaper Base Repair Costs (For Now)
As discussed, the changes were only really to make repairing un-enchanted diamond tools not insanely cheap. Having the cost also scale with how many diamonds it cost kind of made sense, but having a flat cost aside from Leggings and Chestplate works when the enchantments should make the main difference.

If we kept it the same as before, then repairing a Eff5Ub3 or ProtIVUb3 Chestplate would be twice as expensive, and as agreed that is far out from our original balance in line with how much a city will be making. This makes it so that repairing anything isn't completely free, while still making it so that the enchantments make up the majority of the cost increase.